Move _make_pseudo_rgb to _helpers
