+++ plantcv/plantcv/hyperspectral/__init__.py
@@ -1,4 +1,3 @@
-from plantcv.plantcv.hyperspectral.read_data import _make_pseudo_rgb
 from plantcv.plantcv.hyperspectral.read_data import read_data
 from plantcv.plantcv.hyperspectral.extract_wavelength import extract_wavelength
 from plantcv.plantcv.hyperspectral.calibrate import calibrate
@@ -8,5 +7,5 @@ from plantcv.plantcv.hyperspectral.rot90 import rot90
 from plantcv.plantcv.hyperspectral.write_data import write_data
 
 # add new functions to end of lists
-           "_make_pseudo_rgb", "extract_wavelength", "_avg_reflectance", "_inverse_covariance",
+__all__ = ["read_data", "calibrate", "extract_wavelength", "_avg_reflectance", "_inverse_covariance",
            "rot90", "write_data"]
