+++ plantcv/plantcv/hyperspectral/calibrate.py
@@ -5,7 +5,7 @@ import numpy as np
 from plantcv.plantcv import params
 from plantcv.plantcv._debug import _debug
 from plantcv.plantcv import Spectral_data
-from plantcv.plantcv.hyperspectral.read_data import _make_pseudo_rgb
+from plantcv.plantcv._helpers import _make_pseudo_rgb
 
 
 def calibrate(raw_data, white_reference, dark_reference):
